fix: make minimal package changes to come in line with modern header/footer
header and footer have been updated in edx-internal, so this brings
package.json in line with those versions and makes minimal changes to
stop conflicts. Paragon is still on v19 but frontend-platform had to
be updated to v2.

experiments trying to jump paragon as far as 19.25 failed in npm build,
but I've included a snapshot update script command stolen
from elsewhere in edx that was useful to update tests for changing icons